+++ sdk/nodejs/heatmapChart.ts
@@ -6,6 +6,55 @@ import * as inputs from "./types/input";
 import * as outputs from "./types/output";
 import * as utilities from "./utilities";
 
+/**
+ * This chart type displays the specified plot in a heatmap fashion. This format is similar to the [Infrastructure Navigator](https://signalfx-product-docs.readthedocs-hosted.com/en/latest/built-in-content/infra-nav.html#infra), with squares representing each source for the selected metric, and the color of each square representing the value range of the metric.
+ * 
+ * ## Example Usage
+ * 
+ * 
+ * 
+ * ```typescript
+ * import * as pulumi from "@pulumi/pulumi";
+ * import * as signalfx from "@pulumi/signalfx";
+ * 
+ * const myheatmapchart0 = new signalfx.HeatmapChart("myheatmapchart0", {
+ *     colorRange: {
+ *         color: "#ff0000",
+ *         maxValue: 100,
+ *         minValue: 0,
+ *     },
+ *     colorScales: [
+ *         // You can only use one of colorRange or color_scale!
+ *         {
+ *             color: "green",
+ *             gte: 99,
+ *         },
+ *         {
+ *             color: "yellow",
+ *             gte: 95,
+ *             lt: 99, // This ensures we cover the range 95-99
+ *         },
+ *         {
+ *             color: "red",
+ *             lt: 95,
+ *         },
+ *     ],
+ *     description: "Very cool Heatmap",
+ *     disableSampling: true,
+ *     groupBies: [
+ *         "hostname",
+ *         "host",
+ *     ],
+ *     hideTimestamp: true,
+ *     programText: `myfilters = filter("clusterName", "prod") and filter("role", "search")
+ * data("cpu.total.idle", filter=myfilters).publish()
+ * `,
+ *     sortBy: "+host",
+ * });
+ * ```
+ *
+ * > This content is derived from https://github.com/terraform-providers/terraform-provider-signalfx/blob/master/website/docs/r/heatmap_chart.html.markdown.
+ */
 export class HeatmapChart extends pulumi.CustomResource {
     /**
      * Get an existing HeatmapChart resource's state with the given name, ID, and optional extra