+++ sdk/nodejs/orgToken.ts
@@ -83,8 +83,7 @@ export class OrgToken extends pulumi.CustomResource {
      */
     public readonly name!: pulumi.Output<string>;
     /**
-     * List of strings specifying where notifications will be sent when an incident occurs. See
-     * https://developers.signalfx.com/v2/docs/detector-model#notifications-models for more info
+     * Where to send notifications about this token's limits. Please consult the `Notification Format` laid out in detectors.
      */
     public readonly notifications!: pulumi.Output<string[] | undefined>;
     /**
@@ -157,8 +156,7 @@ export interface OrgTokenState {
      */
     readonly name?: pulumi.Input<string>;
     /**
-     * List of strings specifying where notifications will be sent when an incident occurs. See
-     * https://developers.signalfx.com/v2/docs/detector-model#notifications-models for more info
+     * Where to send notifications about this token's limits. Please consult the `Notification Format` laid out in detectors.
      */
     readonly notifications?: pulumi.Input<pulumi.Input<string>[]>;
     /**
@@ -192,8 +190,7 @@ export interface OrgTokenArgs {
      */
     readonly name?: pulumi.Input<string>;
     /**
-     * List of strings specifying where notifications will be sent when an incident occurs. See
-     * https://developers.signalfx.com/v2/docs/detector-model#notifications-models for more info
+     * Where to send notifications about this token's limits. Please consult the `Notification Format` laid out in detectors.
      */
     readonly notifications?: pulumi.Input<pulumi.Input<string>[]>;
 }