+++ sdk/nodejs/types/input.ts
@@ -398,6 +398,14 @@ export interface DetectorVizOption {
     valueUnit?: pulumi.Input<string>;
 }
 
+export interface GetAwsServicesService {
+    name: string;
+}
+
+export interface GetAzureServicesService {
+    name: string;
+}
+
 export interface HeatmapChartColorRange {
     /**
      * The color range to use. Hex values are not supported here. Must be either gray, blue, light_blue, navy, dark_orange, orange, dark_yellow, magenta, cerise, pink, violet, purple, gray_blue, dark_green, green, aquamarine, red, yellow, vivid_yellow, light_green, lime_green.