+++ sdk/python/pulumi_signalfx/list_chart.py
@@ -110,6 +110,49 @@ class ListChart(pulumi.CustomResource):
         The name of each value in the chart reflects the name of the plot and any associated dimensions. We recommend you click the Pencil icon and give the plot a meaningful name, as in plot B below. Otherwise, just the raw metric name will be displayed on the chart, as in plot A below.
 
 
+        ## Example Usage
+
+
+
+        ```python
+        import pulumi
+        import pulumi_signalfx as signalfx
+
+        mylistchart0 = signalfx.ListChart("mylistchart0",
+            color_by="Metric",
+            description="Very cool List Chart",
+            disable_sampling=True,
+            legend_options_fields=[
+                {
+                    "enabled": False,
+                    "property": "collector",
+                },
+                {
+                    "enabled": True,
+                    "property": "cluster_name",
+                },
+                {
+                    "enabled": True,
+                    "property": "role",
+                },
+                {
+                    "enabled": False,
+                    "property": "collector",
+                },
+                {
+                    "enabled": False,
+                    "property": "host",
+                },
+            ],
+            max_delay=2,
+            max_precision=2,
+            program_text=\"\"\"myfilters = filter("cluster_name", "prod") and filter("role", "search")
+        data("cpu.total.idle", filter=myfilters).publish()
+
+        \"\"\",
+            refresh_interval=1,
+            sort_by="-value")
+        ```
 
 
         :param str resource_name: The name of the resource.