+++ sdk/python/pulumi_signalfx/org_token.py
@@ -44,8 +44,7 @@ class OrgToken(pulumi.CustomResource):
     """
     notifications: pulumi.Output[list]
     """
-    List of strings specifying where notifications will be sent when an incident occurs. See
-    https://developers.signalfx.com/v2/docs/detector-model#notifications-models for more info
+    Where to send notifications about this token's limits. Please consult the `Notification Format` laid out in detectors.
     """
     secret: pulumi.Output[str]
     """
@@ -55,6 +54,28 @@ class OrgToken(pulumi.CustomResource):
         """
         Manage SignalFx org tokens.
 
+        ## Example Usage
+
+
+
+        ```python
+        import pulumi
+        import pulumi_signalfx as signalfx
+
+        myteamkey0 = signalfx.OrgToken("myteamkey0",
+            description="My team's rad key",
+            host_or_usage_limits={
+                "containerLimit": 200,
+                "containerNotificationThreshold": 180,
+                "customMetricsLimit": 1000,
+                "customMetricsNotificationThreshold": 900,
+                "highResMetricsLimit": 1000,
+                "highResMetricsNotificationThreshold": 900,
+                "hostLimit": 100,
+                "hostNotificationThreshold": 90,
+            },
+            notifications=["Email,foo-alerts@bar.com"])
+        ```
 
 
         :param str resource_name: The name of the resource.
@@ -64,8 +85,7 @@ class OrgToken(pulumi.CustomResource):
         :param pulumi.Input[dict] dpm_limits: Specify DPM-based limits for this token.
         :param pulumi.Input[dict] host_or_usage_limits: Specify Usage-based limits for this token.
         :param pulumi.Input[str] name: Name of the token.
-        :param pulumi.Input[list] notifications: List of strings specifying where notifications will be sent when an incident occurs. See
-               https://developers.signalfx.com/v2/docs/detector-model#notifications-models for more info
+        :param pulumi.Input[list] notifications: Where to send notifications about this token's limits. Please consult the `Notification Format` laid out in detectors.
 
         The **dpm_limits** object supports the following:
 
@@ -127,8 +147,7 @@ class OrgToken(pulumi.CustomResource):
         :param pulumi.Input[dict] dpm_limits: Specify DPM-based limits for this token.
         :param pulumi.Input[dict] host_or_usage_limits: Specify Usage-based limits for this token.
         :param pulumi.Input[str] name: Name of the token.
-        :param pulumi.Input[list] notifications: List of strings specifying where notifications will be sent when an incident occurs. See
-               https://developers.signalfx.com/v2/docs/detector-model#notifications-models for more info
+        :param pulumi.Input[list] notifications: Where to send notifications about this token's limits. Please consult the `Notification Format` laid out in detectors.
         :param pulumi.Input[str] secret: The secret token created by the API. You cannot set this value.
 
         The **dpm_limits** object supports the following: