+++ sdk/python/pulumi_signalfx/pagerduty/integration.py
@@ -28,6 +28,18 @@ class Integration(pulumi.CustomResource):
 
         > **NOTE** When managing integrations you'll need to use an admin token to authenticate the SignalFx provider. Otherwise you'll receive a 4xx error.
 
+        ## Example Usage
+
+
+
+        ```python
+        import pulumi
+        import pulumi_signalfx as signalfx
+
+        pagerduty_myteam = signalfx.pagerduty.Integration("pagerdutyMyteam",
+            api_key="1234567890",
+            enabled=True)
+        ```
 
 
         :param str resource_name: The name of the resource.