+++ CHANGELOG.md
@@ -2,7 +2,8 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
-* Upgrade to v4.19.7 of the SignalFx Terraform Provider
+* Upgrade to v4.20.1 of the SignalFx Terraform Provider
+* Upgrade to pulumi-terraform-bridge v2.3.1
 
 ---